Replace hardcoded car info with db table

+++ section-2/step-02/src/main/java/com/carmanagement/agentic/agents/CarConditionFeedbackAgent.java
@@ -32,7 +32,7 @@ public interface CarConditionFeedbackAgent {
             String carMake,
             String carModel,
             Integer carYear,
-            Integer carNumber,
+            Long carNumber,
             String carCondition,
             String rentalFeedback,
             String carWashFeedback);
