+++ section-2/step-03/src/main/java/com/carmanagement/agentic/agents/CarWashFeedbackAgent.java
@@ -35,7 +35,7 @@ public interface CarWashFeedbackAgent {
             String carMake,
             String carModel,
             Integer carYear,
-            Integer carNumber,
+            Long carNumber,
             String carCondition,
             String rentalFeedback,
             String carWashFeedback,
