+++ section-2/step-04/multi-agent-system/src/main/java/com/carmanagement/agentic/workflow/CarProcessingWorkflow.java
@@ -24,7 +24,7 @@ public interface CarProcessingWorkflow {
             String carMake,
             String carModel,
             Integer carYear,
-            Integer carNumber,
+            Long carNumber,
             String carCondition,
             String rentalFeedback,
             String carWashFeedback,
